testing/web-platform/tests/old-tests/submission/Infraware/Session_History/contents/Session_History/history_pushstate.html

<!DOCTYPE HTML>
<html>
 <head>
  <title>Session History</title>
  <script src="../../resources/testharness.js"></script>
  <script src="../../resources/testharnessreport.js"></script>
 </head>
 <body>
	<p>
		<h3>history_pushState</h3>
	</p>
	
	<hr>

	<div id="log"></div>

	<script>
		
		try {
			window.history.pushState(1, document.title, '?x=1');

			var state;
			state = window.history.state;
			test(function() {
				assert_equals(state, 1, "pushState method is failed.");
			});
		} catch (e) {
			test(function() {
				assert_unreached("call of pushState method is failed.");
			});
		}

	</script>
 </body>
</html>
